+++ CHANGELOG.md
@@ -1,6 +1,10 @@
+## [Unreleased]
+### Added
+- Adding improved functions for sorting, pagination and views. @bestit-tkellner
+
 ## [3.2.1] - 2017-07-13
 ### Added
-- Adding functionality for form labels for (l)enums @kellner
+- Adding functionality for form labels for (l)enums @bestit-tkellner
 
 ## [3.2.0] - 2017-05-22
 ### Added